Upgrades should work now (hopefully).  You may need to use "--force" on RPM systems if it's the same build number (OMI will always be the same build number, since it doesn't have a build number).
     Some curl changes to use the "easy" interface instead of the callback function.

     There's now a file at /opt/omi/etc/dsc/dsc.conf that contains three keyvalue pairs.

Some additional changes to fix a crash, and allow for a user to not specify a cipher list (in case their version of curl does not support that option. I encountered this on CentOS 7 with my seemingly broken libcurl. Once I rebuilt and reinstalled it, I never saw that issue again. But this change will prevent a customer from simply not being able to ever use the pull client).
Made a change so that we start omiserver after the omi package is installed.
Added support for WMF 5 pull server meta configurations without partial configuration. We should no longer crash here :)