Avoid `make reg` error when OMI is a symlink
In the special case of the `$(OMI)` directory being a symlink and DSC
being configured with `--local` (that is, a typical developer setup),
the `make reg` step would fail before registering `libdsccore` because
it attempted to copy a file from OMI to OMI, and in this case they're
the same directory.

Only attempt the copy if the OMI directory is not a symlink.