Don't upload the files which are greater than specified size (#388)
* Return MaxContentReturnable in the Inventory Record in the Content Field.

The Content Field is not used and will never be used to send content. Re purpose this field to return the size of content that would have been returned instead of the content. This value will be used in the calling code to know what was the value of requested size of content.

* Fix Content to Return MaxContentReturnable

* Fix Contents to Return MaxContentsReturnable